feat: only render staging icons on review page (#824)
* feat: only render staging icons on review page

* fix: constant boolean ternary

* fix: add minus icon
